+++ include/append_buffer.h
@@ -1,0 +1,10 @@
+struct abuf
+{
+    char *b;
+    int len;
+};
+
+#define ABUF_INIT \
+    {             \
+        NULL, 0   \
+    }
