+++ kilo.c
@@ -1,42 +1,5 @@
-/*** includes ***/
-
-#include <stdio.h>
-#include <stdlib.h>
-#include <errno.h>
-#include <ctype.h>
-#include <string.h>
-#include <sys/ioctl.h>
-#include <unistd.h>
-#include <termios.h>
-
-/*** data ***/
-struct editorConfig
-{
-    int cx, cy;
-    int screenrows;
-    int screencols;
-    struct termios orig_termios;
-};
-
-struct editorConfig E;
-
-/*** defines ***/
-
-#define KILO_VERSION "0.0.1"
-#define CTRL_KEY(k) ((k) & 0x1f)
-
-enum editorKey
-{
-    ARROW_LEFT = 1000,
-    ARROW_RIGHT,
-    ARROW_UP,
-    ARROW_DOWN,
-    PAGE_UP,
-    PAGE_DOWN,
-    HOME_KEY,
-    END_KEY,
-    DEL_KEY,
-};
+#include "./include/data.h"
+#include "./include/append_buffer.h"
 
 /*** terminal ***/
 
@@ -216,17 +179,6 @@ int getWindowSize(int *rows, int *cols)
 }
 
 /*** append buffer ***/
-
-struct abuf
-{
-    char *b;
-    int len;
-};
-
-#define ABUF_INIT \
-    {             \
-        NULL, 0   \
-    }
 
 void abAppend(struct abuf *ab, const char *s, int len)
 {
